sua kieu du lieu migration

+++ database/migrations/2024_09_25_065008_create_attributes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('attributes', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->tinyInteger('status')->default(1)->comment('1:active; 2: inactive');
+            $table->integer('status')->default(1)->comment('1:active; 2: inactive');
             $table->timestamps();
             $table->softDeletes();
         });
