+++ database/migrations/2024_09_25_072728_create_pages_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('slug')->unique();
             $table->text('content');
-            $table->tinyInteger('status')->default(1)->comment('1: active, 2: inactive');
+            $table->integer('status')->default(1)->comment('1: active, 2: inactive');
             $table->softDeletes();
             $table->timestamps();
         });
